refactor(llm): move markdown sanitizer
Move `strip_markdown_fence()` into `tests/test_sanitized.py` and
add tests to verify fenced Markdown output is unwrapped correctly for
Telegram rendering.
